requirements.txt
Flask==3.0.0
gunicorn==21.2.0
firecrawl
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
ratelimit==2.2.1
beautifulsoup4==4.12.2
lxml==4.9.3